Move HostConcatResultUtil out of unshimmed classes (#5614)
Don't use ai.rapidsai.cudf package for spark-rapids Scala classes. Otherwise it is going to be loaded by conventional classloader and fail to load referenced classes out of the shimmed areas.
- Move the class 
- add a smoking test to prevent this sort of regressions in premerge

Closes #5513. 

Depends on rapidsai/cudf#10949
